fix(scan): magic-link verify page must dispatch via Workflow, not legacy in-process pipeline
The /scan/verify/[token] page was orphaned during the #615/#618 migrations
to Cloudflare Workflows. The API endpoint at /api/scan/verify was correctly
updated to dispatch via the SCAN_WORKFLOW_SERVICE binding (the durable
ss-scan-workflow Worker), but the Astro page kept calling the legacy
runDiagnosticScan inline inside ctx.waitUntil. Worker isolate budget kills
the waitUntil task mid-pipeline (4 Claude calls ~30s+ wall clock), so
prospects verified their email, the page rendered "your scan is running,"
and the report email never arrived.

Live tail captured the smoking gun:
  GET https://smd.services/scan/verify/REDACTED - Ok
    (warn) waitUntil() tasks did not complete within the allowed time
           after invocation end and have been cancelled.

Fix: extract handleVerify + runFallback into src/lib/scan/verify-handler.ts.
Both surfaces (API endpoint and Astro page) now call the same helper —
single source of truth for verify+dispatch. Production goes through the
service binding to ss-scan-workflow; dev/test still falls back to inline
ctx.waitUntil so astro dev and vitest keep working.

Regression-guard tests added: page must import handleVerify, must NOT
import runDiagnosticScan, must NOT call markScanVerified directly. These
are static-source assertions — same pattern as the existing scan-endpoints
suite — so the next time someone refactors verify they can't accidentally
re-orphan the page.
