+++ src/pages/scan/verify/[token].astro
@@ -3,11 +3,13 @@
  * Magic-link verification landing page (#598).
  *
  * The prospect clicks the link in their inbox and lands here. The page
- * server-side calls the verify API which:
- *   - Hashes the inbound token, looks up the scan_request row.
- *   - Marks it verified and kicks off the diagnostic pipeline via
- *     ctx.waitUntil (so the response returns instantly).
- *   - Returns a status string we render below.
+ * delegates verify+dispatch to `lib/scan/verify-handler.ts`, which is the
+ * same code path `/api/scan/verify` uses — production dispatches the
+ * scan as a Cloudflare Workflow on the dedicated `ss-scan-workflow`
+ * Worker (durable, retries, no isolate-budget limit). Before this shared
+ * helper the page ran the legacy inline pipeline via `ctx.waitUntil`,
+ * which the Worker isolate budget killed mid-scan, so prospects verified
+ * but never received a report.
  *
  * The page never echoes specific findings — those only ever arrive via
  * email per the scoping doc Bar #6 (anti-competitor-intel). The page
@@ -18,10 +20,7 @@
 import Base from '../../../layouts/Base.astro'
 import Nav from '../../../components/Nav.astro'
 import Footer from '../../../components/Footer.astro'
-import { env } from 'cloudflare:workers'
-import { hashScanToken, isScanTokenFresh } from '../../../lib/scan/tokens'
-import { getScanRequestByTokenHash, markScanVerified } from '../../../lib/db/scan-requests'
-import { runDiagnosticScan } from '../../../lib/diagnostic'
+import { handleVerify } from '../../../lib/scan/verify-handler'
 
 const { token } = Astro.params
 
@@ -33,45 +32,29 @@ type RenderState =
   | { kind: 'invalid' }
   | { kind: 'failed' }
 
-let state: RenderState = { kind: 'invalid' }
+const result = await handleVerify(typeof token === 'string' ? token : '', Astro.locals)
 
-if (typeof token === 'string' && token.length > 0) {
-  const hash = await hashScanToken(token)
-  const row = await getScanRequestByTokenHash(env.DB, hash)
-  if (!row) {
-    state = { kind: 'invalid' }
-  } else if (row.scan_status === 'completed') {
-    state = { kind: 'already_completed', domain: row.domain }
-  } else if (row.scan_status === 'thin_footprint') {
-    state = { kind: 'thin_footprint', domain: row.domain }
-  } else if (row.scan_status === 'failed') {
-    state = { kind: 'failed' }
-  } else if (!isScanTokenFresh(row.created_at)) {
+let state: RenderState
+switch (result.status) {
+  case 'verified':
+    state = { kind: 'verified', domain: result.domain ?? '' }
+    break
+  case 'already_completed':
+    state = { kind: 'already_completed', domain: result.domain ?? '' }
+    break
+  case 'thin_footprint':
+    state = { kind: 'thin_footprint', domain: result.domain ?? '' }
+    break
+  case 'expired':
     state = { kind: 'expired' }
-  } else {
-    await markScanVerified(env.DB, row.id)
-    state = { kind: 'verified', domain: row.domain }
-
-    // Fire the pipeline via ctx.waitUntil so it survives the response
-    // closing. Captain memory: never inline-await LLM pipelines.
-    const ctx = Astro.locals.cfContext
-    const scanPromise = runDiagnosticScan(
-      {
-        DB: env.DB,
-        ANTHROPIC_API_KEY: env.ANTHROPIC_API_KEY,
-        GOOGLE_PLACES_API_KEY: env.GOOGLE_PLACES_API_KEY,
-        OUTSCRAPER_API_KEY: env.OUTSCRAPER_API_KEY,
-        RESEND_API_KEY: env.RESEND_API_KEY,
-        APP_BASE_URL: env.APP_BASE_URL,
-      },
-      row.id
-    )
-    if (ctx?.waitUntil) {
-      ctx.waitUntil(scanPromise)
-    } else {
-      scanPromise.catch((err) => console.error('[scan/verify] dev-fallback scan threw:', err))
-    }
-  }
+    break
+  case 'failed':
+    state = { kind: 'failed' }
+    break
+  case 'invalid_token':
+  default:
+    state = { kind: 'invalid' }
+    break
 }
 ---
